pyproject.toml
[project]
name = "map-tracker"
version = "0.1.0"
description = "游戏地图跟点助手"
requires-python = ">=3.9"
dependencies = [
    "opencv-python",
    "numpy",
    "mss",
    "Pillow",
    "requests",
    "flask>=2.0",
    "flask-socketio>=5.3",
    "python-socketio>=5.11",
    "orjson>=3.10",
    "flask-compress>=1.15",
    "gevent>=25.0",
    "gevent-websocket>=0.10",
]

[project.optional-dependencies]
ai = [
    "torch",
    "kornia",
]

[project.scripts]
map-tracker-web  = "run_web:main"
map-tracker-sift = "run_sift:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
norecursedirs = [
    ".git",
    ".venv",
    "_debug",
    "node_modules",
]